+++ feature/home/src/main/java/com/ilsangtech/ilsang/feature/home/component/LargeRewardQuestCard.kt
@@ -1,4 +1,4 @@
-package com.ilsangtech.ilsang.feature.home.home.component
+package com.ilsangtech.ilsang.feature.home.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
